requirements.txt
Django>=3.1.7,<3.2.0
djangorestframework>=3.12.2,<3.13.0
pillow>=8.1.2,<8.2.0